linkCreateAccount/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.4
Django==3.2.14
django-rest-knox==4.2.0
djangorestframework==3.13.1
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.3.0

django-cors-headers==3.5.0
psycopg2-binary==2.8.5
pandas==1.1.1
scikit-learn==0.23.2
joblib==0.16.0
gunicorn==20.0.4
Pillow==9.0.1